requirements.txt
streamlit
passwordstrength
translate
Faker
pandas
numpy
opencv-python-headless
Pillow
xlsxwriter
chardet
baidu-aip
streamlit-cropper
jieba
wordcloud
matplotlib
streamlit_echarts
cryptography